+++ component/build.gradle
@@ -13,6 +13,7 @@ dependencies {
     implementation project(':alert-common')
     implementation project(':alert-database')
     implementation project(':alert-database-job')
+    implementation project(':api-authentication')
     implementation project(':api-common-model')
     implementation project(':api-descriptor')
     implementation project(':api-environment')
